Add API for /v3/access_rules_config
Add Flask resources for access rules config. This exposes listing all
configured access rules or filtering by service type, which users can use to
be informed about what rules are valid. It does not expose the ability to
check a rule against the configured list, which is only used internally.
This API is deliberately unprotected.

bp whitelist-extension-for-app-creds

Change-Id: I9e25087b2ddb6e76e39eb62893378678a5cbc03c