Update introduction of external services doc
This document is written for other developers working on OpenStack.
Its goal is to describe various concepts in a way that other
developers, who may not be familiar with keystone, can digest and use
effectively in their projects.

The introductory paragraph was phrased as if the v3 API just became a
thing, when it has actually been around for a long time. It also
eluded to underlying implementation details by mentioning paste
pipelines, which we no longer use.

This commit updates the introduction to be more relevant to the state
of things today.

Change-Id: I79564dc99fd65a5609bd548d12a0413ca3ee6b2a