Switch federation check jobs to opensuse
Ubuntu Xenial is nearing its end of life, so it's not ideal to keep
testing on it and the QA team is driving an effort to move everyone away
from it[1]. However, our non-voting federation jobs rely on the
Shibboleth service provider Apache module, which does not work on Ubuntu
Bionic[2]. Since the keystone devstack plugin now has support for CentOS
and openSUSE, let's transition away from Ubuntu for this testing.

The reason to choose openSUSE rather then CentOS is that the Shibboleth
service provider package for CentOS relies on an external repository and
is not included in either the base CentOS distro or EPEL. The
shibboleth-sp package for openSUSE is included in the base distro
packages.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2019-February/003129.html
[2] https://bugs.launchpad.net/ubuntu/+source/shibboleth-sp2/+bug/1776489

Change-Id: I89a8af9c45ff513c854c048880854990a6d12278