Pin Werkzeug in lower-constraints
upper-constraints pins Werkzeug to a known working version. The
lower-constraints job uses a different constraints file, the
locally-maintained lower-constraints.txt, which didn't pin Werkzeug and
resulted in the job breaking on the most recent 0.15.0 Werkzeug release.
This change pins Werkzeug to match the known working version in
upper-constraints. If and when upper-constraints raises its cap, we can
address the breakage on the regular unit test jobs.

Change-Id: I926e79e34072505f9edc4879b1d9237a3b85accd