Ignore Stein-specific release notes
We merged and backported a lot of fixes for the Stein RC. Since they
landed in master first, reno renders them in the "unreleased" page, but
since they are about to be released for Stein it's not appropriate to
leave them there, so this change explicitly ignores them. We can also
now remove old ignored release notes.

Change-Id: Iad8bf7e03dd9672add9f01794c9b3e0ba4257831