Revert "Add manager support for app cred access rules"
This reverts commit 37fc2b9120538e001daba7093d0295e5f65b811d.

In the Train PTG[1] we agreed to defer the access rules config part of
this feature until we had some kind of traceability or discoverability
for APIs. For simplicity of review, this patch reverts the access rules
addition to the app cred manager so that we can reimplement it in a way
that doesn't require using the access_rules_config API provider.

[1] https://etherpad.openstack.org/p/keystone-train-ptg-application-credentials

Change-Id: I65ac52b8730221562391adc8b0dbccd22ea79b16