Add exercises for intern applicants
At the Train PTG, we discussed[1] what we could do to continue to
participate in open source internship programs like Outreachy while
keeping the high volume of applicants manageable. Outreachy requires
applicants to submit a contribution to the project they are applying
for, but we rarely have sufficient numbers of low-hanging-fruit tasks to
assign to all of them to allow all of them to make useful contributions.
Instead, we propose to give them exercises that most likely will not be
merged into keystone itself. These exercises are much more difficult
than typo fix tasks, and can be assigned to multiple people without
concern for applicants stepping on each others' toes. They can also help
new contributors get familiar with the architecture of keystone and the
development workflow, and encourages them to interact with the team.

[1] https://etherpad.openstack.org/p/keystone-train-ptg-outreachy-brainstorm

Change-Id: I615b1c029db59a1f9d8548dc0a80faa5c4150f2a