Allow JsonBlob to accommodate SQL NULL result sets
SQLAlchemy passes a SQL NULL result value, which in Python comes
back as None, to the TypeDecorator.process_result_value method
for a datatype that is handling that column.  Python's json.loads()
function does not accommodate Python None directly.  A NULL
value may be returned for a JsonBlob expression either if SELECTing
from a table where the value is NULL, as there are several
NULLable columns in Keystone's model which use this type, or
if the column is linked to a SQL result such as from an outer join
that does not contain the value.

The reason this bug has not been seen in Keystone is because the only
occurrence of SQL NULL for this column type has been in an ORM
"joinedload()" LEFT OUTER JOIN expression, and the ORM does not
fetch the actual columns when the identity is NULL.  However, upcoming
changes in SQLAlchemy include that the lower level row object
will be firing off these result handlers up front, rather than
deferring them for fetch, which has illustrated this issue.

The patch includes a test suite exercising the JsonBlob and
DateTimeInt types completely, as there do not seem to be tests that
target these constructs individually.

Change-Id: Ib1448fa08d7106fd55e18a0f92a705011bc3ce2b