Update api-ref location
The api documentation is now published on docs.openstack.org instead
of developer.openstack.org. Update all links that are changed to the
new location.

This does not relationship URIs since "these links do not resolve to
anything valid, but exist to show a relationship."

Note that redirects will be set up as well but let's point now to the
new location.

For details, see:
http://lists.openstack.org/pipermail/openstack-discuss/2019-July/007828.html

Change-Id: I6efdf375bc8e1e5ca2b113337002d6178180a1e1