Fix websso auth loop
In bf67b3c8[1] we introduced an error whereby trying to use WebSSO to
authenticate with one protocol would fail with an uncaught HTTP 404 if
an IdP was found that did not use that protocol. This patch fixes the
issue by ensuring that during the search for an IdP that matches the
given protocol, we ignore invalid IdPs.

This is tested by the existing WebSSOTests unit test class simply by
inserting a dummy IdP and protocol combination into the test data during
setup, since the problem arises when the protocol you are *not* trying
to authenticate with is indexed first in the database.

Since the breaking change was not released yet, this bugfix does not
need a release note.

[1] https://review.opendev.org/637305

Change-Id: Id423f8a304abffbe0c7814ab2ab4458e6a403bb1
Closes-bug: #1838592