Fixing 404 URLs for Rocky
It would be preferable to backport
https://review.opendev.org/#/c/675667/ but the difficulty (merge
conflicts) in backporting it was considered too high for a small
change due to the dramatic difference in the doc structure from
rocky to latest (Train). Instead, it is proposed to merge
these nit link changes into stable only.

This patch was proposed directly to Rocky in conjunction
with these two backports:
1. https://review.opendev.org/676955
2. https://review.opendev.org/683945

Change-Id: I17ae42d5e14fa09f44f430ffa01c01c33a96f57e