Implement system admin for trusts API
This change enables a system admin to delete trusts. Previously, only
the trustor or the is_admin admin could delete a trust. This changes
makes the trusts API more useful to system administrators who need to
clean up trusts and makes the API consistent with others.

This does not enable system admins to create trusts. A trust can only be
scoped to a project, so creating one is inherently a project-scoped
action. If trusts later gain the ability to be scoped to the system or
domains, we can add those scopes to the create_trust scope_types.

Change-Id: Idf13b862f345388bb2372609787947eb43d7ba75
Closes-bug: #1818846
Closes-bug: #1818850
Related-Bug: #968696