Add notifications for deleting app creds by user
Without this patch, when an individual application credential is deleted
it emits a notification, but when all are deleted for a user there is no
notification (although this is only triggered when a user is deleted or
disabled or has a change in role assignments, all of which generate
their own notifications). This patch ensures audit notifications are
generated for every application credential that gets deleted.

Change-Id: I7c820931585802c7afa53727623ac05adee56248