Implement system reader for implied roles
This change updates the implied roles policies to understand the system
reader role, and includes tests for system reader and system member. A
follow-on change will add support for system admin. For the time being,
we're deferring adding support for domain users to create implied roles
but may add it in the future.

Change-Id: I71f80eed5cd689505c7fd07f133e74f0e8edf66a
Partial-bug: #1805371