Override tox job timeouts
Currently our unit tests frequently hit the 40 minute timeout due to the
inefficiency of the protection unit tests. As a temporary workaround,
this change overrides the zuul tox jobs with a 60 minute timeout.
Overriding the values of jobs that are defined elsewhere in a project
template isn't exactly documented but it supposedly works[1].

[1] http://eavesdrop.openstack.org/irclogs/%23openstack-infra/%23openstack-infra.2019-09-06.log.html#t2019-09-06T23:35:57

Change-Id: I0b9e0fbbd1760fe2d55935592f4349f337801209