Increase tox job timeouts to 90 minutes
With the gate being extra-busy due to feature freeze week, we seem to be
having even more of a noisy neighbor problem and jobs are still often
hitting the 60 minute timeout. This change increases the timeout from 60
to 90 minutes. This is not an indication that things are normal and fine
but should hopefully alleviate pressure until we're in a position to
merge a more satisfactory workaround and fix.

Change-Id: I171eccc2dd46c26ada74def36523c6f7f29be868