Use correct repo for initial version check
Use the actual repo for the initial version check of the db version when
performing `db_sync --check`. Previously we always used the Legacy Repo
as the repo path was not populated in the `get_init_version` function.

This allowed the case where the new expand/contract/migrate repos
matched the legacy repo version number, the `db_sync --check` command
would assume that everything was properly up to date regardless of the
version for the expand/contract/migrate repositories. This was most
noticable in the case of a db not under version control.

Change-Id: Id5e6f74a5ed96fdf17e2c241466897974aa5a218
closes-bug: #1844157