Overhaul the RBAC documentation for administrators
The administrator guide has a section dedicated to RBAC and policy.
This guide doesn't include any information about the work we've been
doing since Pike to evolve keystone's default policies.

This commit updates the documentation to describe what comes with
keystone RBAC out-of-the-box for operators. It also reduces the amount
of documentation that was duplicated with what oslo.policy already has
written down.

Change-Id: Ia870a4353b8c6b0725615e65bd1c32745c9975b3