Fix K2K auth flow diagram
The keystone-to-keystone auth flow diagram contained a minor error that
could lead to confusion. When a client POSTs a SAMLResponse to the PAOS
URL of the keystone Service Provider, the response is not automatically
a keystone token but an HTTP redirect that indicates the session is
successfully authenticated. The client then must ignore the Location
header of the redirect and instead start the token request process with
keystone's federated authentication endpoint. This change adds the extra
step to the diagram to help clarify the process.

Change-Id: I5c256388c2247eba4e559eb94cc9fb1bcd42444a